Move app to shorter name and add django app config
