+++ src/saml2_pro_auth/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "saml2_pro_auth.apps.Saml2ProAuthConfig"
