+++ src/saml2_pro_auth/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class Saml2ProAuthConfig(AppConfig):
+    name = "saml2_pro_auth"
+    verbose_name = "Django SAML2 Pro Authentication"
