+++ tests/settings.py
@@ -5,4 +5,4 @@ DATABASES = {"default": {"ENGINE": "django.db.backends.sqlite3"}}
 ALLOWED_HOSTS = ["*"]
 
 SAML_ROUTE = "/sso/saml"
-ROOT_URLCONF = "django_saml2_pro_auth.urls"
+ROOT_URLCONF = "saml2_pro_auth.urls"
