+++ tox.ini
@@ -36,4 +36,4 @@ addopts = --ds=tests.settings
 django_find_project = false
 
 [coverage:run]
-source=src/django_saml2_pro_auth
+source=src/saml2_pro_auth
